# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clinic",
        "description": "API پروژه ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/provider/{id}": {
            "get": {
                "tags": [
                    "Provider"
                ],
                "summary": "دریافت ارائه دهنده با شناسه",
                "operationId": "getProviderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "دکتر هاشمی"
                                                },
                                                "speciality": {
                                                    "type": "string",
                                                    "example": "متخصص حلق و بینی"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Provider not found"
                    }
                }
            }
        },
        "/api/provider/{id}/available-slots?date=": {
            "get": {
                "tags": [
                    "Provider"
                ],
                "summary": "دریافت تاریخ‌های در دسترس بودن ارائه‌دهنده بر اساس تاریخ دریافتی",
                "operationId": "getProviderDatesByDay",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Day of week ",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1404-10-11"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "provider_id": {
                                                        "type": "string",
                                                        "example": "دکتر حسین محمدی"
                                                    },
                                                    "speciality": {
                                                        "type": "string",
                                                        "example": "دکتر عمومی"
                                                    },
                                                    "day_of_week": {
                                                        "type": "string",
                                                        "example": "1404-10-11"
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "09:00 AM"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "17:00 PM"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "دریافت ویزیت های فعال",
                "operationId": "getActiveServices",
                "responses": {
                    "200": {
                        "description": "Active services found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "ویزیت متخصص قلب و عروق"
                                                    },
                                                    "duration_minutes": {
                                                        "type": "string",
                                                        "example": "30 Min"
                                                    },
                                                    "price": {
                                                        "type": "string",
                                                        "example": "200,000"
                                                    },
                                                    "is_active": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No active services",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "اتمام ویزیت"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "Bearer {token}"
            }
        }
    }
}